requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
Django==5.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
redis==5.0.8
sqlparse==0.5.1
typing_extensions==4.12.2
